Fix: dev환경에는 트래킹 요청대신 콘솔로그 출력되도록 수정
