+++ src/hoc/useAnalytics.tsx
@@ -1,23 +1,7 @@
 import mixpanel from 'mixpanel-browser';
 import React, { createContext, useContext, useEffect, useState } from 'react';
 
-export const eventNames = [
-  'Tool_Click',
-  'User',
-  'Tool_Category_Click',
-  'Integrated_Search_Enter',
-  'Banner_Click',
-  'Toggle_Click',
-  'Sorting_Click',
-  'Tool_Detail_Index_Click',
-  'Recommendation_Tool_Click',
-  'Community_Click',
-  'Post_Click',
-  'Signup_Click',
-  'Login_State',
-  'Signout_Click',
-] as const;
-type EventName = (typeof eventNames)[number];
+import { EventName } from '@constants/event';
 
 type AnalyticsContextProps = {
   trackEvent: <T extends Record<string, unknown>>(eventName: EventName, eventProperties?: T) => void;
@@ -27,10 +11,15 @@ const AnalyticsContext = createContext<AnalyticsContextProps | undefined>(undefi
 
 const AnalyticsProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [isReady, setIsReady] = useState(false);
+  const isDev = import.meta.env.MODE === 'development';
 
   useEffect(() => {
+    if (isDev) {
+      setIsReady(true);
+      return;
+    }
     mixpanel.init(import.meta.env.VITE_MIXPANEL_KEY, {
-      debug: import.meta.env.MODE === 'development',
+      debug: isDev,
       loaded: () => setIsReady(true),
     });
   }, []);
@@ -38,6 +27,10 @@ const AnalyticsProvider: React.FC<{ children: React.ReactNode }> = ({ children }
   const trackEvent = <T extends Record<string, unknown>>(eventName: EventName, eventProperties?: T) => {
     if (!isReady) {
       console.warn('Mixpanel이 아직 준비되지 않았습니다');
+      return;
+    }
+    if (isDev) {
+      console.log('[TRACKING]', eventName, eventProperties);
       return;
     }
     mixpanel.track(eventName, eventProperties);
